fixes to checkout and purchaase history stuff
