+++ src/bookshop/index.js
@@ -1,6 +1,5 @@
 import React from 'react';
 import { BrowserRouter, Route, Switch } from 'react-router-dom';
-import { useDispatch } from 'react-redux';
 
 import Footer from './components/Footer';
 import Header from './components/Header';
@@ -10,17 +9,12 @@ import Nav from './components/Nav';
 import Cart from './Cart';
 import Featured from './Featured';
 
-import { storeHistory } from '../actions';
-
 import '../App.css';
 import 'bootstrap/dist/css/bootstrap.css';
 
 
 const Bookshop = ({ firebase }) => {
 
-  const dispatch = useDispatch();
-  dispatch(storeHistory());
-  
   return (
     <BrowserRouter>
       <Nav firebase={firebase} />
